Add prefix paths for Exherbo's filesystem layout
Exherbo uses a special filesystem layout which separates libraries and
binaries based on what CHOST they were compiled for. /usr/lib, on Exherbo,
is just a symlink to whatever the host machine's correct lib folder would be,
which would break if running for i686-pc-linux-gnu on a x86_64-pc-linux-gnu host.